+++ src/reporter/command/format-command.ts
@@ -1,0 +1,8 @@
+import { Command, FormattedCommand } from './interfaces';
+import { CommandFormatter } from './command-formatter';
+
+export default function formatCommand (command: Command, result: unknown): FormattedCommand {
+    const commandFormatter = new CommandFormatter(command, result);
+
+    return commandFormatter.format();
+}
